+++ driver/src/test/scala/org/mongodb/scala/ReadPreferenceSpec.scala
@@ -26,7 +26,7 @@ import org.scalatest.{FlatSpec, Matchers}
 
 class ReadPreferenceSpec extends FlatSpec with Matchers {
 
-  val duration = Duration("1 sec")
+  val duration = Duration("95 sec")
 
   "ReadPreference" should "have the same methods as the wrapped ReadPreference" in {
     val wrapped = classOf[com.mongodb.ReadPreference].getDeclaredMethods.filter(f => isStatic(f.getModifiers)).map(_.getName).toSet
@@ -45,25 +45,25 @@ class ReadPreferenceSpec extends FlatSpec with Matchers {
     readPreference shouldBe com.mongodb.ReadPreference.primaryPreferred()
 
     val readPreference1 = ReadPreference.primaryPreferred(duration)
-    readPreference1 shouldBe com.mongodb.ReadPreference.primaryPreferred(1, SECONDS)
+    readPreference1 shouldBe com.mongodb.ReadPreference.primaryPreferred(95, SECONDS)
 
     val readPreference2 = ReadPreference.primaryPreferred(TagSet())
     readPreference2 shouldBe com.mongodb.ReadPreference.primaryPreferred(TagSet())
 
     val readPreference3 = ReadPreference.primaryPreferred(TagSet(), duration)
-    readPreference3 shouldBe com.mongodb.ReadPreference.primaryPreferred(TagSet(), 1, SECONDS)
+    readPreference3 shouldBe com.mongodb.ReadPreference.primaryPreferred(TagSet(), 95, SECONDS)
 
     val readPreference4 = ReadPreference.primaryPreferred(TagSet(Tag("name", "value")))
     readPreference4 shouldBe com.mongodb.ReadPreference.primaryPreferred(TagSet(Tag("name", "value")))
 
     val readPreference5 = ReadPreference.primaryPreferred(TagSet(Tag("name", "value")), duration)
-    readPreference5 shouldBe com.mongodb.ReadPreference.primaryPreferred(TagSet(Tag("name", "value")), 1, SECONDS)
+    readPreference5 shouldBe com.mongodb.ReadPreference.primaryPreferred(TagSet(Tag("name", "value")), 95, SECONDS)
 
     val readPreference6 = ReadPreference.primaryPreferred(List(TagSet(List(Tag("name", "value")))))
     readPreference6 shouldBe com.mongodb.ReadPreference.primaryPreferred(List(TagSet(List(Tag("name", "value")))).asJava)
 
     val readPreference7 = ReadPreference.primaryPreferred(List(TagSet(List(Tag("name", "value")))), duration)
-    readPreference7 shouldBe com.mongodb.ReadPreference.primaryPreferred(List(TagSet(List(Tag("name", "value")))).asJava, 1, SECONDS)
+    readPreference7 shouldBe com.mongodb.ReadPreference.primaryPreferred(List(TagSet(List(Tag("name", "value")))).asJava, 95, SECONDS)
   }
 
   it should "return the correct secondary based ReadPreferences" in {
@@ -71,25 +71,25 @@ class ReadPreferenceSpec extends FlatSpec with Matchers {
     readPreference shouldBe com.mongodb.ReadPreference.secondary()
 
     val readPreference1 = ReadPreference.secondary(duration)
-    readPreference1 shouldBe com.mongodb.ReadPreference.secondary(1, SECONDS)
+    readPreference1 shouldBe com.mongodb.ReadPreference.secondary(95, SECONDS)
 
     val readPreference2 = ReadPreference.secondary(TagSet())
     readPreference2 shouldBe com.mongodb.ReadPreference.secondary(TagSet())
 
     val readPreference3 = ReadPreference.secondary(TagSet(), duration)
-    readPreference3 shouldBe com.mongodb.ReadPreference.secondary(TagSet(), 1, SECONDS)
+    readPreference3 shouldBe com.mongodb.ReadPreference.secondary(TagSet(), 95, SECONDS)
 
     val readPreference4 = ReadPreference.secondary(TagSet(Tag("name", "value")))
     readPreference4 shouldBe com.mongodb.ReadPreference.secondary(TagSet(Tag("name", "value")))
 
     val readPreference5 = ReadPreference.secondary(TagSet(Tag("name", "value")), duration)
-    readPreference5 shouldBe com.mongodb.ReadPreference.secondary(TagSet(Tag("name", "value")), 1, SECONDS)
+    readPreference5 shouldBe com.mongodb.ReadPreference.secondary(TagSet(Tag("name", "value")), 95, SECONDS)
 
     val readPreference6 = ReadPreference.secondary(List(TagSet(List(Tag("name", "value")))))
     readPreference6 shouldBe com.mongodb.ReadPreference.secondary(List(TagSet(List(Tag("name", "value")))).asJava)
 
     val readPreference7 = ReadPreference.secondary(List(TagSet(List(Tag("name", "value")))), duration)
-    readPreference7 shouldBe com.mongodb.ReadPreference.secondary(List(TagSet(List(Tag("name", "value")))).asJava, 1, SECONDS)
+    readPreference7 shouldBe com.mongodb.ReadPreference.secondary(List(TagSet(List(Tag("name", "value")))).asJava, 95, SECONDS)
   }
 
   it should "return the correct secondaryPreferred based ReadPreferences" in {
@@ -97,25 +97,25 @@ class ReadPreferenceSpec extends FlatSpec with Matchers {
     readPreference shouldBe com.mongodb.ReadPreference.secondaryPreferred()
 
     val readPreference1 = ReadPreference.secondaryPreferred(duration)
-    readPreference1 shouldBe com.mongodb.ReadPreference.secondaryPreferred(1, SECONDS)
+    readPreference1 shouldBe com.mongodb.ReadPreference.secondaryPreferred(95, SECONDS)
 
     val readPreference2 = ReadPreference.secondaryPreferred(TagSet())
     readPreference2 shouldBe com.mongodb.ReadPreference.secondaryPreferred(TagSet())
 
     val readPreference3 = ReadPreference.secondaryPreferred(TagSet(), duration)
-    readPreference3 shouldBe com.mongodb.ReadPreference.secondaryPreferred(TagSet(), 1, SECONDS)
+    readPreference3 shouldBe com.mongodb.ReadPreference.secondaryPreferred(TagSet(), 95, SECONDS)
 
     val readPreference4 = ReadPreference.secondaryPreferred(TagSet(Tag("name", "value")))
     readPreference4 shouldBe com.mongodb.ReadPreference.secondaryPreferred(TagSet(Tag("name", "value")))
 
     val readPreference5 = ReadPreference.secondaryPreferred(TagSet(Tag("name", "value")), duration)
-    readPreference5 shouldBe com.mongodb.ReadPreference.secondaryPreferred(TagSet(Tag("name", "value")), 1, SECONDS)
+    readPreference5 shouldBe com.mongodb.ReadPreference.secondaryPreferred(TagSet(Tag("name", "value")), 95, SECONDS)
 
     val readPreference6 = ReadPreference.secondaryPreferred(List(TagSet(List(Tag("name", "value")))))
     readPreference6 shouldBe com.mongodb.ReadPreference.secondaryPreferred(List(TagSet(List(Tag("name", "value")))).asJava)
 
     val readPreference7 = ReadPreference.secondaryPreferred(List(TagSet(List(Tag("name", "value")))), duration)
-    readPreference7 shouldBe com.mongodb.ReadPreference.secondaryPreferred(List(TagSet(List(Tag("name", "value")))).asJava, 1, SECONDS)
+    readPreference7 shouldBe com.mongodb.ReadPreference.secondaryPreferred(List(TagSet(List(Tag("name", "value")))).asJava, 95, SECONDS)
   }
 
   it should "return the correct nearest based ReadPreferences" in {
@@ -123,19 +123,19 @@ class ReadPreferenceSpec extends FlatSpec with Matchers {
     readPreference shouldBe com.mongodb.ReadPreference.nearest()
 
     val readPreference1 = ReadPreference.nearest(duration)
-    readPreference1 shouldBe com.mongodb.ReadPreference.nearest(1, SECONDS)
+    readPreference1 shouldBe com.mongodb.ReadPreference.nearest(95, SECONDS)
 
     val readPreference2 = ReadPreference.nearest(TagSet(Tag("name", "value")))
     readPreference2 shouldBe com.mongodb.ReadPreference.nearest(TagSet(Tag("name", "value")))
 
     val readPreference3 = ReadPreference.nearest(TagSet(Tag("name", "value")), duration)
-    readPreference3 shouldBe com.mongodb.ReadPreference.nearest(TagSet(Tag("name", "value")), 1, SECONDS)
+    readPreference3 shouldBe com.mongodb.ReadPreference.nearest(TagSet(Tag("name", "value")), 95, SECONDS)
 
     val readPreference4 = ReadPreference.nearest(List(TagSet(List(Tag("name", "value")))))
     readPreference4 shouldBe com.mongodb.ReadPreference.nearest(List(TagSet(List(Tag("name", "value")))).asJava)
 
     val readPreference5 = ReadPreference.nearest(List(TagSet(List(Tag("name", "value")))), duration)
-    readPreference5 shouldBe com.mongodb.ReadPreference.nearest(List(TagSet(List(Tag("name", "value")))).asJava, 1, SECONDS)
+    readPreference5 shouldBe com.mongodb.ReadPreference.nearest(List(TagSet(List(Tag("name", "value")))).asJava, 95, SECONDS)
   }
 
   it should "return the correct ReadPreference for valueOf" in {
@@ -146,7 +146,7 @@ class ReadPreferenceSpec extends FlatSpec with Matchers {
     readPreference2 shouldBe com.mongodb.ReadPreference.primaryPreferred(List(TagSet(Tag("name", "value"))).asJava)
 
     val readPreference3 = ReadPreference.valueOf("PrimaryPreferred", List(TagSet(Tag("name", "value"))), duration)
-    readPreference3 shouldBe com.mongodb.ReadPreference.primaryPreferred(List(TagSet(Tag("name", "value"))).asJava, 1, SECONDS)
+    readPreference3 shouldBe com.mongodb.ReadPreference.primaryPreferred(List(TagSet(Tag("name", "value"))).asJava, 95, SECONDS)
   }
 
 }